Bug fix in agg_det_slider
Bug fix in agg_det_slider, wherein the median filter was only applied once. This resulted in jagged slider thresholding that did not match previous implementation. Gaussian filter is now applied 8 times. This remains inconsistent with the previous software in that the Gaussian filter was applied every time the slider button was clicked. This is an inconsistent approach, such that the more standard 8 times was selected here.
+ Clarifications to naming convention
+ Updates to where defaults for rolling ball transformation are specified (now it only occurs in the rolling_ball.m function, unifying the defaults across multiple methods)